+++ tests/image_test.cc
@@ -82,6 +82,25 @@ TEST(Image, animator) {
   widget_destroy((w));
 }
 
+TEST(Image, animator2) {
+  widget_t* w = (image_create(NULL, 0, 0, 400, 300));
+
+  ASSERT_EQ(widget_create_animator(w, ";move(y_from=0, y_to=128);;rotation(to=6.28, yoyo_times=0, duration=1000, easing=sin_out)"), RET_OK);
+  ASSERT_EQ(widget_start_animator(w, "move"), RET_OK);
+  ASSERT_EQ(widget_pause_animator(w, "move"), RET_OK);
+  ASSERT_EQ(widget_find_animator(w, "move") != NULL, TRUE);
+  ASSERT_EQ(widget_destroy_animator(w, "move"), RET_OK);
+  ASSERT_EQ(widget_find_animator(w, "move") == NULL, TRUE);
+
+  ASSERT_EQ(widget_start_animator(w, "rotation"), RET_OK);
+  ASSERT_EQ(widget_pause_animator(w, "rotation"), RET_OK);
+  ASSERT_EQ(widget_find_animator(w, "rotation") != NULL, TRUE);
+  ASSERT_EQ(widget_destroy_animator(w, "rotation"), RET_OK);
+  ASSERT_EQ(widget_find_animator(w, "rotation") == NULL, TRUE);
+
+  widget_destroy((w));
+}
+
 TEST(Image, cast) {
   widget_t* w = image_create(NULL, 0, 0, 400, 300);