+++ docs/changes.md
@@ -4,6 +4,7 @@
   * 增加EMITTER\_ENABLE/EMITTER\_DISABLE
   * 完善text edit。
   * 完善widget\_set\_tr\_text(感谢俊圣提供补丁)
+  * 完善widget\_create\_animator(感谢俊圣提供补丁)
 
 2022/05/19
   * 完善func\_str